+++ src/modules/chrome/accordion/anchorButton.js
@@ -3,7 +3,7 @@ import createLi from '../../common/cElement/createLi';
 import insertAfterParent from './insertAfterParent';
 import insertElement from '../../common/insertElement';
 import insertElementAfter from '../../common/insertElementAfter';
-import jQueryDialog from '../jQueryDialog';
+import jQueryDialog from '../jQueryDialog/jQueryDialog';
 import onclick from '../../common/onclick';
 import partial from '../../common/partial';
 import sendEvent from '../../analytics/sendEvent';
