--- src/modules/chrome/helperMenu.js
@@ -1,128 +1,0 @@
-import createDiv from '../common/cElement/createDiv';
-import draggable from '../common/dragStart';
-import getElementById from '../common/getElement';
-import getText from '../common/getText';
-import getValue from '../system/getValue';
-import hasClass from '../common/hasClass';
-import insertElement from '../common/insertElement';
-import insertHtmlBeforeEnd from '../common/insertHtmlBeforeEnd';
-import isFunction from '../common/isFunction';
-import jQueryDialog from './jQueryDialog';
-import jQueryPresent from '../common/jQueryPresent';
-import once from '../common/once';
-import onclick from '../common/onclick';
-import sendEvent from '../analytics/sendEvent';
-import {
-  injectAuctionSearch,
-  injectBuffLog,
-  injectFindBuffs,
-  injectFindOther,
-  injectFsBoxContent,
-  injectMonsterLog,
-  injectNotepadShowLogs,
-  injectOnlinePlayers,
-  injectQuickLinkManager,
-  injectRecipeManager,
-  insertQuickExtract,
-  insertQuickWear,
-} from './pageSwitcher/loader';
-import { newGuildLogUrl, notepadBlankUrl, playerIdUrl } from '../support/constants';
-
-const helperMenuBlob = '<div class="column"><h3>Character</h3><ul>'
-  + '<li><span class="fshLink">Buff Log</span></li>'
-  + '<li><span class="fshLink">Combat Log</span></li>'
-  + '<li><span class="fshLink">Creature Log</span></li>'
-  + '<li><span class="fshLink">Recipe Manager</span></li>'
-  + '<li><span class="fshLink">Quick Links</span></li>'
-  + `<li><a href="${notepadBlankUrl}invmanagernew">Inventory Manager</a></li>`
-  + '</ul><h3>Actions</h3><ul>'
-  + '<li><span class="fshLink">Find Buffs</span></li>'
-  + '<li><span class="fshLink">Find Other</span></li>'
-  + '<li><span class="fshLink">Online Players</span></li>'
-  + '<li><span class="fshLink">AH Quick Search</span></li>'
-  + '</ul><h3>Guild</h3><ul>'
-  + `<li><a href="${notepadBlankUrl}guildinvmgr">Guild Inventory</a></li>`
-  + `<li><a href="${newGuildLogUrl}">New Guild Log</a></li>`
-  + '</ul><h3>Extra</h3><ul>'
-  + '<li><span class="fshLink">Quick Extract</span></li>'
-  + '<li><span class="fshLink">Quick Wear</span></li>'
-  + '<li><span class="fshLink">FS Box Log</span></li>'
-  + '</ul><h3>FSH developer quick links</h3><ul>'
-  + '<li><span class="a-reply" target_player="PointyHair">PM</span> '
-  + `<a href="${playerIdUrl}1963510">PointyHair</a></li>`
-  + '</ul></div>';
-
-function toggleMenu(evt) {
-  if (evt.target.id !== 'helperMenu') { return; }
-  const menu = evt.target.children[0];
-  menu.classList.toggle('showMenuDiv');
-}
-
-const functionLookup = {
-  'Buff Log': injectBuffLog,
-  'Combat Log': injectNotepadShowLogs,
-  'Creature Log': injectMonsterLog,
-  'Recipe Manager': injectRecipeManager,
-  'Quick Links': injectQuickLinkManager,
-  'Find Buffs': injectFindBuffs,
-  'Find Other': injectFindOther,
-  'Online Players': injectOnlinePlayers,
-  'AH Quick Search': injectAuctionSearch,
-  'Quick Extract': insertQuickExtract,
-  'Quick Wear': insertQuickWear,
-  'FS Box Log': injectFsBoxContent,
-};
-
-function callHelperFunction(evt) {
-  const functionPath = getText(evt.target);
-  const fn = functionLookup[functionPath];
-  if (jQueryPresent() && isFunction(fn)) {
-    sendEvent('helperMenu', functionPath);
-    jQueryDialog(fn);
-  }
-}
-
-function eventHandler(evt) {
-  if (hasClass('fshLink', evt.target)) {
-    callHelperFunction(evt);
-    return;
-  }
-  if (hasClass('a-reply', evt.target)) {
-    window.openQuickMsgDialog(evt.target.getAttribute('target_player'));
-  }
-}
-
-function showHelperMenu(evt) {
-  const helperMenu = evt.target;
-  const helperMenuDiv = createDiv({
-    id: 'helperMenuDiv',
-    className: 'helperMenuDiv fshInnerBg',
-  });
-  insertHtmlBeforeEnd(helperMenuDiv, helperMenuBlob);
-  insertElement(helperMenu, helperMenuDiv);
-  onclick(helperMenu, toggleMenu);
-  onclick(helperMenuDiv, eventHandler);
-}
-
-function haveNode() {
-  const helperMenu = createDiv({
-    id: 'helperMenu',
-    className: 'helperMenu',
-    innerHTML: 'Helper&nbsp;Menu',
-  });
-  if (getValue('keepHelperMenuOnScreen')) {
-    helperMenu.classList.add('fshFixed');
-  }
-  once(helperMenu, 'mouseenter', showHelperMenu);
-  if (getValue('draggableHelperMenu')) {
-    helperMenu.classList.add('fshMove');
-    draggable(helperMenu);
-  }
-  insertElement(document.body, helperMenu);
-}
-
-export default function injectHelperMenu() {
-  // don't put all the menu code here (but call if clicked) to minimize lag
-  const node = getElementById('statbar-container');
-  if (node) { haveNode(); }
-}
