+++ src/modules/chrome/jQueryDialog/jQueryDialog.css
@@ -1,0 +1,1 @@
+#content {max-width:600px}
