+++ src/modules/profile/ifSelf/quickWearLink.js
@@ -3,7 +3,7 @@ import createSpan from '../../common/cElement/createSpan';
 import insertElement from '../../common/insertElement';
 import { insertQuickWear } from '../../chrome/pageSwitcher/loader';
 import insertTextBeforeEnd from '../../common/insertTextBeforeEnd';
-import jQueryDialog from '../../chrome/jQueryDialog';
+import jQueryDialog from '../../chrome/jQueryDialog/jQueryDialog';
 import onclick from '../../common/onclick';
 import querySelector from '../../common/querySelector';
 import sendEvent from '../../analytics/sendEvent';
